initial restructuring
