+++ constants/screen.py
@@ -1,0 +1,6 @@
+from .base import SPRITE_SIZE
+
+WIDTH = SPRITE_SIZE * 15
+HEIGHT = SPRITE_SIZE * 10
+
+TITLE = "Dungeons and Directories"
