+++ game_objects/__init__.py
@@ -1,0 +1,9 @@
+from .player import Player
+from .level import Level
+from .game_object import GameObject
+
+__all__ = [
+        "Player",
+        "Level",
+        "GameObject",
+]
