+++ managers/__init__.py
@@ -1,0 +1,7 @@
+from .game_window import MyWindow
+from .physics import Physics
+
+__all__ = [
+        "Physics",
+        "MyWindow",
+]
