+++ managers/physics.py
